Add a function to build the entire `scriptletsGlobals` (with debug flag) and `deAmpEnabled` scripts we inject with scriptlets. iOS will pass in `canDebug` as false as it's currently unsupported / unused on iOS.
