+++ components/cosmetic_filters/common/BUILD.gn
@@ -6,7 +6,12 @@
 import("//mojo/public/tools/bindings/mojom.gni")
 
 source_set("common") {
-  sources = [ "scriptlet_constants.h" ]
+  sources = [
+    "scriptlet_constants.cc",
+    "scriptlet_constants.h",
+  ]
+
+  deps = [ "//third_party/abseil-cpp:absl" ]
 }
 
 mojom("mojom") {
