+++ ios/brave-ios/Sources/Brave/Frontend/Browser/UserScripts/ScriptFactory.swift
@@ -167,14 +167,14 @@ class ScriptFactory {
     case .engineScript(let configuration):
       let source = [
         "(function(){",
-        // This map is used by some of uBlock Origin's resources
-        AdblockService.scriptletGlobalsScript(),
-        // This boolean is used by a script injected by cosmetic filters and enables that script via this boolean
-        // The script is found here: https://github.com/brave/adblock-resources/blob/master/resources/de-amp.js
+        // Defines `scriptletGlobals`, used by some of uBlock Origin's resources,
+        // and `deAmpEnabled`, used by a script injected by cosmetic filters.
+        // The de-amp script is found here:
+        // https://github.com/brave/adblock-resources/blob/master/resources/de-amp.js
         // - Note: This script is only a smaller part (1 of 3) of de-amping:
         // The second part is handled by an inected script that redirects amp pages to their canonical links
         // The third part is handled by debouncing amp links and handled by debouncing rules
-        configuration.isDeAMPEnabled ? "const deAmpEnabled = true;" : "",
+        AdblockService.scriptletGlobalsScript(withDeAmpEnabled: configuration.isDeAMPEnabled),
         configuration.source,
         "})();",
       ].joined(separator: "\n")
